+++ CHANGELOG.md
@@ -7,7 +7,7 @@ All notable changes to this project will be documented in this file. This projec
 ### What's Changes
 * Added `nextPeriodTime` option to the config for the `thermostatSetpointStatus` option `TemporaryHold`.
 
-**Full Changelog**: https://github.com/donavanbecker/homebridge-honeywell-home/compare/v11.1.2...v11.2.0
+**Full Changelog**: https://github.com/donavanbecker/homebridge-honeywell-home/compare/v11.2.0...v11.3.0
 
 ## [Version 11.2.0](https://github.com/donavanbecker/homebridge-honeywell-home/releases/tag/v11.2.0) (2021-12-18)